inform about songs not in library
Before, song ID lookup by URL was integrated in the SQL statement to
insert the found ID into the playlist. When the URL was not found, this
simply put an empty entry into the list.

Splicing lookup and insertion apart enables to avoid this by checking
whether some id has been found and informing the user if the song is
missing from the library.

It not even cost of a whole separate statement, as two queries now do
the same thing that one did before. There is just some program logic in
between.