utilize git template for source array
Not only does this standardize the source retrieval, it also configures
the correct package relations to its non-git variant and sets up a
proper pkgver function.

Because the project does not have any version, the pkgver is set to
revision 0 so that the version numbers generated always adjust up, not
causing any unnecessary update indication.

As the PKGBUILD now has to be processed with makepkg-expanded, its
intermediate relics are filed as irrelevant to tracking.